moved GetLightClientHeightSafely
