+++ x/stakeibc/keeper/interchainaccounts.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/bech32"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/gogoproto/proto"
-	"github.com/spf13/cast"
 
 	"github.com/Stride-Labs/stride/v18/utils"
 	icacallbackstypes "github.com/Stride-Labs/stride/v18/x/icacallbacks/types"
@@ -311,28 +310,4 @@ func (k Keeper) SubmitICATxWithoutCallback(
 	}
 
 	return nil
-}
-
-func (k Keeper) GetLightClientHeightSafely(ctx sdk.Context, connectionID string) (uint64, error) {
-	// get light client's latest height
-	conn, found := k.IBCKeeper.ConnectionKeeper.GetConnection(ctx, connectionID)
-	if !found {
-		errMsg := fmt.Sprintf("invalid connection id, %s not found", connectionID)
-		k.Logger(ctx).Error(errMsg)
-		return 0, fmt.Errorf(errMsg)
-	}
-	clientState, found := k.IBCKeeper.ClientKeeper.GetClientState(ctx, conn.ClientId)
-	if !found {
-		errMsg := fmt.Sprintf("client id %s not found for connection %s", conn.ClientId, connectionID)
-		k.Logger(ctx).Error(errMsg)
-		return 0, fmt.Errorf(errMsg)
-	} else {
-		latestHeightHostZone, err := cast.ToUint64E(clientState.GetLatestHeight().GetRevisionHeight())
-		if err != nil {
-			errMsg := fmt.Sprintf("error casting latest height to int64: %s", err.Error())
-			k.Logger(ctx).Error(errMsg)
-			return 0, fmt.Errorf(errMsg)
-		}
-		return latestHeightHostZone, nil
-	}
 }
